branitz_energy_decision_ai_street_final/requirements.txt
# Core validation dependencies
jsonschema>=4.19
pytest>=7.4

# Code quality tools
ruff>=0.5
black>=24.0

# Optional dependencies (for full functionality)
pandas>=2.0
numpy>=1.24
geopandas>=0.13
shapely>=2.0
pandapipes>=0.8
pandapower>=2.14
pyyaml>=6.0
openpyxl>=3.1
scikit-learn>=1.3
lightgbm>=4.0
pyarrow>=12.0  # For parquet files

# Google ADK for enhanced multi-agent system
adk>=0.1.0
google-generativeai>=0.3.0

# ADK-specific dependencies
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Enhanced multi-agent system dependencies
psutil>=5.9.0  # For performance monitoring
questionary>=1.10.0  # For interactive prompts
jinja2>=3.1.0  # For template rendering
markdown>=3.4.0  # For documentation rendering

# Development dependencies
pytest-cov>=4.0 